fix setup structure folders
