--- frontend/app/(root)/Loginpage.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-
-export default function LoginPage() 
-{
-    return (
-        <main>
-            <div>Login Page</div>
-            <div>
-                Logout
-            </div>
-        </main>
-    )
-}
