+++ frontend/app/Pages/HomePage/Components/Content.tsx
