+++ frontend/app/Pages/HomePage/Components/Hero.tsx
