+++ frontend/app/Pages/HomePage/Components/HotFeatures.tsx
